graph-neo4j/scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "neo4j>=5.28.1",
    "python-dotenv>=1.0.1",
]
